fix: Move accountEntityType to separate file for react-refresh
ESLint react-refresh/only-export-components rule prevents exporting
utility functions from component files. Move to dedicated module.
