+++ frontend/src/features/positions/pages/detail.tsx
@@ -7,7 +7,7 @@ import { TimeDisplay } from '@/shared/time-display'
 import { QualityLadderBadge } from '@/features/positions/components/quality-ladder-badge'
 import { DirectionBadge } from '@/shared/direction-badge'
 import { EntityLink, Breadcrumbs } from '@/shared'
-import { accountEntityType } from '@/shared/entity-link'
+import { accountEntityType } from '@/shared/account-entity-type'
 import { Skeleton } from '@/components/ui/skeleton'
 import { usePositionLogDetail } from '../hooks'
 import type { FinancialPositionLog, TransactionLogEntry } from './index'
